move the destructuring tests to their own file
macro-test.test.ts evaluates top-level macro imports at module load,
which can trip a pre-existing !m_topGCOwnedDataScope debug assertion and
abort before any test runs. The destructuring tests are self-contained
subprocess tests, so isolating them keeps local debug verification
deterministic.
